pyproject.toml
[project]
name = "key-rotator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "keyring>=25.0",
    "secretstorage>=3.3; sys_platform == 'linux'",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "fastapi>=0.136.0",
    "uvicorn[standard]>=0.46.0",
]

[project.scripts]
key-rotator = "rotator.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["rotator"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyinstaller>=6.20.0",
]
